@@ -3,10 +3,8 @@
  */
 import * as bedrock from '@bedrock/core';
 import * as helpers from './helpers.js';
-import {createRequire} from 'node:module';
 import {keystores} from '@bedrock/kms';
-const require = createRequire(import.meta.url);
-const {KeystoreConfigStorage} = require('@digitalbazaar/webkms-switch');
+import {KeystoreConfigStorage} from '@digitalbazaar/webkms-switch';
 
 const {util: {BedrockError}} = bedrock;
 
