@@ -6,7 +6,6 @@ import * as brZCapStorage from '@bedrock/zcap-storage';
 import * as helpers from './helpers.js';
 import * as middleware from './middleware.js';
 import {asyncHandler} from '@bedrock/express';
-import {createRequire} from 'node:module';
 import {createValidateMiddleware as validate} from '@bedrock/validation';
 import {
   defaultModuleManager as moduleManager,
@@ -19,8 +18,7 @@ import {
   postRevocationBody
 } from '../schemas/bedrock-kms-http.js';
 import {reportOperationUsage, SERVICE_TYPE} from './metering.js';
-const require = createRequire(import.meta.url);
-const {generateRandom} = require('@digitalbazaar/webkms-switch');
+import {generateRandom} from '@digitalbazaar/webkms-switch';
 
 const {config, util: {BedrockError}} = bedrock;
 const {cors} = middleware;
