@@ -6,7 +6,6 @@ import * as brZCapStorage from '@bedrock/zcap-storage';
 import * as helpers from './helpers.js';
 import {asyncHandler} from '@bedrock/express';
 import {BedrockKeystoreConfigStorage} from './BedrockKeystoreConfigStorage.js';
-import {createRequire} from 'node:module';
 import {default as _cors} from 'cors';
 import {
   defaultDocumentLoader as documentLoader,
@@ -14,17 +13,16 @@ import {
 } from '@bedrock/kms';
 import {logger} from './logger.js';
 import {reportOperationUsage} from './metering.js';
-const require = createRequire(import.meta.url);
-const {
-  authorizeZcapInvocation: _authorizeZcapInvocation,
-  authorizeZcapRevocation: _authorizeZcapRevocation
-} = require('@digitalbazaar/ezcap-express');
-const {createOperationMiddleware} = require('@digitalbazaar/webkms-switch');
-const {CryptoLD} = require('crypto-ld');
-const {Ed25519Signature2020} =
-  require('@digitalbazaar/ed25519-signature-2020');
-const {Ed25519VerificationKey2020} =
-  require('@digitalbazaar/ed25519-verification-key-2020');
+import {
+  authorizeZcapInvocation as _authorizeZcapInvocation,
+  authorizeZcapRevocation as _authorizeZcapRevocation
+} from '@digitalbazaar/ezcap-express';
+import {createOperationMiddleware} from '@digitalbazaar/webkms-switch';
+import {CryptoLD} from 'crypto-ld';
+import {Ed25519Signature2020} from '@digitalbazaar/ed25519-signature-2020';
+import {
+  Ed25519VerificationKey2020
+} from '@digitalbazaar/ed25519-verification-key-2020';
 
 const {config, util: {BedrockError}} = bedrock;
 
