@@ -4,15 +4,14 @@
 import * as bedrock from '@bedrock/core';
 import * as helpers from './helpers.js';
 import {agent} from '@bedrock/https-agent';
-import {createRequire} from 'node:module';
 import {httpClient, DEFAULT_HEADERS} from '@digitalbazaar/http-client';
 import {mockData} from './mock.data.js';
-const require = createRequire(import.meta.url);
-const {CapabilityAgent} = require('@digitalbazaar/webkms-client');
-const {signCapabilityInvocation} = require(
-  '@digitalbazaar/http-signature-zcap-invoke');
-const {ZcapClient} = require('@digitalbazaar/ezcap');
-const {Ed25519Signature2020} = require('@digitalbazaar/ed25519-signature-2020');
+import {CapabilityAgent} from '@digitalbazaar/webkms-client';
+import {
+  signCapabilityInvocation
+} from '@digitalbazaar/http-signature-zcap-invoke';
+import {ZcapClient} from '@digitalbazaar/ezcap';
+import {Ed25519Signature2020} from '@digitalbazaar/ed25519-signature-2020';
 
 describe('bedrock-kms-http API', () => {
   describe('keystores', () => {
