@@ -2,7 +2,7 @@
  * Copyright (c) 2019-2022 Digital Bazaar, Inc. All rights reserved.
  */
 import pMap from 'p-map';
-import uuid from 'uuid-random';
+import {v4 as uuid} from 'uuid';
 import {CapabilityAgent, KeystoreAgent, KmsClient}
   from '@digitalbazaar/webkms-client';
 import * as helpers from './helpers.js';
